Split LLVM op emission categories
